Move socket-related code to server/sockets file
